requirements.txt
black
fastapi
fastapi-sqlalchemy
pydantic
alembic
psycopg2
uvicorn
python-dotenv
pytest
